+++ core/src/main/kotlin/com/jasonernst/kanonproxy/BidirectionalByteChannel.kt
