+++ core/src/main/kotlin/com/jasonernst/kanonproxy/tcp/InitialSequenceNumberGenerator.kt
