+++ core/src/main/kotlin/com/jasonernst/kanonproxy/tcp/TcpState.kt
