+++ core/src/test/resources/logback-test.xml
