+++ settings.gradle.kts
@@ -5,6 +5,9 @@ pluginManagement {
         mavenCentral()
     }
 }
+plugins {
+    id("org.gradle.toolchains.foojay-resolver-convention") version "0.8.0"
+}
 dependencyResolutionManagement {
     repositoriesMode.set(RepositoriesMode.FAIL_ON_PROJECT_REPOS)
     repositories {
@@ -13,3 +16,6 @@ dependencyResolutionManagement {
     }
 }
 rootProject.name = "kanonproxy"
+include("core")
+include("server")
+include("client")
