--- src/main/kotlin/com/jasonernst/kanonproxy/SessionManager.kt
@@ -1,5 +1,0 @@
-package com.jasonernst.kanonproxy
-
-interface SessionManager {
-    fun removeSession(session: Session)
-}
